requirements.txt
flask
flask-socketio
redis
gevent-websocket
uwsgi
gevent
greenlet
itsdangerous
Jinja2
MarkupSafe
python-engineio
python-socketio
redis
six
uWSGI
Werkzeug
eventlet
